refactor: separate app module into submodules for easier code readability and maintainment
